Move getSite up one level
